mma_leads/requirements.txt
Django==5.0.2
djangorestframework==3.15.2
django-cors-headers==4.3.1
Pillow==10.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
